Refactor out SzLoader class
